+++ app/Http/Controllers/Api/OrderController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Order;
 use Illuminate\Http\Request;
 use Validator;
+use App\Permissions\Permission;
 
 class OrderController extends Controller
 {
@@ -17,8 +18,15 @@ class OrderController extends Controller
      */
     public function index()
     {
-        $orders = Order::all();
-        return response()->json($orders);
+        if($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_LIMIT_INDEX)) {
+            $orders = Order::all();
+            return response()->json($orders);
+        }
+
+        return response()->json([
+            'error'=> 'Error',
+            'message' => 'Not Authorized.'
+        ]);
     }
 
     /**
@@ -29,52 +37,59 @@ class OrderController extends Controller
      */
     public function store(Request $request)
     {
-        $input = $request->all();
+        if($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_LIMIT_STORE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'user_name' => 'required|string',
-            'email' => 'required|string',
-            'plan_id' => 'required|string|max:100',
-            'plan_name' => 'required|string',
-            'payment_response' => 'required|string',
-            'payment_status' => 'required|string|max:55',
-            'created_date' => 'required|string',
-            'amount' => 'required|string|max:55',
-            'client_secret' => 'required|string',
-            'fingerprint' => 'required|string',
-            'charge_id' => 'required|string|max:100',
-            'customer_id' => 'required|string|max:100',
-            'currency' => 'required|string|max:50',
-            'exp_month' => 'required|numeric',
-            'exp_year' => 'required|numeric',
-            'card_st_digit' => 'required|numeric',
-            'user_id' => 'required|uuid'
+            $validator = Validator::make($input, [
+                'user_name' => 'required|string',
+                'email' => 'required|string',
+                'plan_id' => 'required|string|max:100',
+                'plan_name' => 'required|string',
+                'payment_response' => 'required|string',
+                'payment_status' => 'required|string|max:55',
+                'created_date' => 'required|string',
+                'amount' => 'required|string|max:55',
+                'client_secret' => 'required|string',
+                'fingerprint' => 'required|string',
+                'charge_id' => 'required|string|max:100',
+                'customer_id' => 'required|string|max:100',
+                'currency' => 'required|string|max:50',
+                'exp_month' => 'required|numeric',
+                'exp_year' => 'required|numeric',
+                'card_st_digit' => 'required|numeric',
+                'user_id' => 'required|uuid'
+            ]);
+            
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
+
+            try {
+                $order = Order::create($input);
+                return response()->json($order);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "Order store error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
+            }
+        }
+
+        return response()->json([
+            'error'=> 'Error',
+            'message' => 'Not Authorized.'
         ]);
-        
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
-            ]);
-        }
-
-        try {
-            $order = Order::create($input);
-            return response()->json($order);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
-            }
-            else{
-                $message = "Order store error";
-            }
-            return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
-            ]);
-        }
     }
 
     /**
@@ -85,7 +100,14 @@ class OrderController extends Controller
      */
     public function show(Order $order)
     {
-        return response()->json($order);
+        if($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_LIMIT_SHOW)) {
+            return response()->json($order);
+        }
+
+        return response()->json([
+            'error'=> 'Error',
+            'message' => 'Not Authorized.'
+        ]);
     }
 
     /**
@@ -97,52 +119,59 @@ class OrderController extends Controller
      */
     public function update(Request $request, Order $order)
     {
-        $input = $request->all();
+        if($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_LIMIT_UPDATE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'user_name' => 'nullable|string',
-            'email' => 'nullable|string',
-            'plan_id' => 'nullable|string|max:100',
-            'plan_name' => 'nullable|string',
-            'payment_response' => 'nullable|string',
-            'payment_status' => 'nullable|string|max:55',
-            'created_date' => 'nullable|string',
-            'amount' => 'nullable|string|max:55',
-            'client_secret' => 'nullable|string',
-            'fingerprint' => 'nullable|string',
-            'charge_id' => 'nullable|string|max:100',
-            'customer_id' => 'nullable|string|max:100',
-            'currency' => 'nullable|string|max:50',
-            'exp_month' => 'nullable|numeric',
-            'exp_year' => 'nullable|numeric',
-            'card_st_digit' => 'nullable|numeric',
-            'user_id' => 'nullable|uuid'
+            $validator = Validator::make($input, [
+                'user_name' => 'nullable|string',
+                'email' => 'nullable|string',
+                'plan_id' => 'nullable|string|max:100',
+                'plan_name' => 'nullable|string',
+                'payment_response' => 'nullable|string',
+                'payment_status' => 'nullable|string|max:55',
+                'created_date' => 'nullable|string',
+                'amount' => 'nullable|string|max:55',
+                'client_secret' => 'nullable|string',
+                'fingerprint' => 'nullable|string',
+                'charge_id' => 'nullable|string|max:100',
+                'customer_id' => 'nullable|string|max:100',
+                'currency' => 'nullable|string|max:50',
+                'exp_month' => 'nullable|numeric',
+                'exp_year' => 'nullable|numeric',
+                'card_st_digit' => 'nullable|numeric',
+                'user_id' => 'nullable|uuid'
+            ]);
+
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
+
+            try {
+                $order->update($input);
+                return response()->json($order);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "Order update error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
+            }
+        }
+
+        return response()->json([
+            'error'=> 'Error',
+            'message' => 'Not Authorized.'
         ]);
-
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
-            ]);
-        }
-
-        try {
-            $order->update($input);
-            return response()->json($order);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
-            }
-            else{
-                $message = "Order update error";
-            }
-            return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
-            ]);
-        }
     }
 
     /**
@@ -153,7 +182,14 @@ class OrderController extends Controller
      */
     public function destroy(Order $order)
     {
-        $order->delete();
-        return response()->json();
+        if($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_LIMIT_DESTROY)) {
+            $order->delete();
+            return response()->json();
+        }
+
+        return response()->json([
+            'error'=> 'Error',
+            'message' => 'Not Authorized.'
+        ]);
     }
 }
