+++ app/Http/Controllers/Api/ResetPasswordController.php
@@ -7,6 +7,7 @@ use App\Models\User;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Validator;
+use App\Permissions\Permission;
 
 class ResetPasswordController extends Controller
 {
@@ -16,40 +17,47 @@ class ResetPasswordController extends Controller
      */
     public function __invoke(Request $request)
     {
-        $input = $request->all();
+        if($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_RESETPASSWORD_INVOKE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'code' => 'required|string|exists:reset_code_passwords',
-            'password' => 'required|string|min:8|regex:/^.*(?=.{3,})(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[\d\x])(?=.*[!$#%]).*$/',
-        ]);
-        
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
+            $validator = Validator::make($input, [
+                'code' => 'required|string|exists:reset_code_passwords',
+                'password' => 'required|string|min:8|regex:/^.*(?=.{3,})(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[\d\x])(?=.*[!$#%]).*$/',
             ]);
+            
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
+
+            $passwordReset = ResetCodePassword::firstWhere('code', $request->code);
+
+            if ($passwordReset->isExpire()) {
+                return response()->json([
+                    'code' => false,
+                    'error' => 'Failed to reset',
+                    'message' => 'Reset code expired',
+                ]);
+            }
+
+            $user = User::firstWhere('email', $passwordReset->email);
+
+            $user->update([
+                'password' => bcrypt($request->password)
+            ]);
+
+            
+            ResetCodePassword::where('code', $request->code)->delete();
+
+            return response()->json(['userInfo' => $user]);
         }
 
-        $passwordReset = ResetCodePassword::firstWhere('code', $request->code);
-
-        if ($passwordReset->isExpire()) {
-            return response()->json([
-                'code' => false,
-                'error' => 'Failed to reset',
-                'message' => 'Reset code expired',
-            ]);
-        }
-
-        $user = User::firstWhere('email', $passwordReset->email);
-
-        $user->update([
-            'password' => bcrypt($request->password)
+        return response()->json([
+            'error'=> 'Error',
+            'message' => 'Not Authorized.'
         ]);
-
-        
-        ResetCodePassword::where('code', $request->code)->delete();
-
-        return response()->json(['userInfo' => $user]);
     }
 }
